+++ src/Area.h
@@ -5,19 +5,19 @@
 namespace Areas
 {
 	typedef struct {
-		long left, top, right, bottom;
+		int left, top, right, bottom;
 	} Area;	
 
-	Area SetArea(long left, long top, long right, long bottom);
-	Area SetMarquee(long left, long top, long right, long bottom);
-	void OffsetArea(Area *a, long dx, long dy);
-	void ScaleArea(Area *a, long scale);
-	Area InsetArea(const Area *area, long dx, long dy);
+	Area SetArea(int left, int top, int right, int bottom);
+	Area SetMarquee(int left, int top, int right, int bottom);
+	void OffsetArea(Area *a, int dx, int dy);
+	void ScaleArea(Area *a, int scale);
+	Area InsetArea(const Area *area, int dx, int dy);
 	Area UniteArea(const Area *a, const Area *b);
 	Area IntersectArea(const Area *a, const Area *b);
-	long SubtractArea(const Area *a, const Area *b, Area part[]); // returns #parts (0..4)
+	int SubtractArea(const Area *a, const Area *b, Area part[]); // returns #parts (0..4)
 	bool SameArea(const Area *a, const Area *b);
-	bool AreaIncludesPoint(const Area *a, long x, long y);
+	bool AreaIncludesPoint(const Area *a, int x, int y);
 	bool AreaIncludesArea(const Area *a, const Area *b);
 	bool AreaIntersectsArea(const Area *a, const Area *b);
 }
